Better SocietyDisplay, better Highway trajectories, and IntelligentPanel improvements
Makes a number of substantial changes to the UI and appearance of the
game. SocietyDisplay now presents a much larger amount of information.
BlobHighways now orient themselves parallel to the MapEdges they occupy.
And IntelligentPanels now keep themselves positioned in the camera's
view and away from the objects they're intended to highlight.